wip - fix s3-yaml circular dep.
