--- treehugger/client.py
@@ -1,6 +1,0 @@
-# -*- coding:utf-8 -*-
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-import boto3
-
-s3_client = boto3.client('s3')
