+++ treehugger/s3.py
@@ -1,13 +1,43 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import boto3
 import six
 from botocore.exceptions import ClientError
 from six.moves.urllib.parse import parse_qs, urlparse
 
 from . import yaml
-from .client import s3_client
 from .messaging import die
+
+s3_client = boto3.client('s3')
+
+
+def _split_s3_filename(s3_key):
+    parsed_url = urlparse(s3_key)
+    if parsed_url.scheme.lower() != 's3':
+        die('Got an unsupported url scheme: {}'.format(parsed_url.scheme))
+    bucket_name = parsed_url.netloc
+    bucket_key = parsed_url.path[1:]
+    version = parse_qs(parsed_url.query)['versionId'][0]
+    return bucket_name, bucket_key, version
+
+
+def read_file_from_s3(filename):
+    bucket_name, bucket_key, version = _split_s3_filename(filename)
+    try:
+        obj = s3_client.get_object(
+            Bucket=bucket_name,
+            Key=bucket_key,
+            VersionId=version,
+        )
+    except ClientError as exc:
+        die('Got "{}" when attempting to fetch key {} version {} from bucket {}'.format(
+            exc.response['Error']['Code'],
+            bucket_key,
+            version,
+            bucket_name,
+        ))
+    return obj['Body'].read()
 
 
 def fetch_s3_content_or_die(bucket_name, key, version):
